bigint for pg

+++ db/migrate/20201019151242_create_servers.rb
@@ -1,7 +1,7 @@
 class CreateServers < ActiveRecord::Migration[6.0]
   def change
     create_table :servers do |t|
-      t.integer :discord_id
+      t.bigint :discord_id
       t.string :name
       t.integer :admin_user_id
       t.timestamps
